move spectraray and nexus importers
