+++ src/elli/__init__.py
@@ -9,7 +9,6 @@ from .solver2x2 import Solver2x2
 from .solver4x4 import *
 from .structure import *
 from .utils import *
-from .spectraray import *
 from .materials_db import *
 
 if sys.version_info.major >= 3 and sys.version_info.minor > 7:
